--- src/Caliburn.Micro.WinAppSdk/Winappsdk/Caliburn.Micro.Platform.rd.xml
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Directives xmlns="http://schemas.microsoft.com/netfx/2013/01/metadata">
-  <Library Name="Caliburn.Micro.Platform.UWP">
-    <Assembly Name="Caliburn.Micro.Platform.UWP" Activate="Required All" Browse="Required All" Serialize="Required All" Dynamic="Required All" />
-  </Library>
-</Directives>
