+++ src/Caliburn.Micro.sln
@@ -17,7 +17,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Caliburn.Micro.Xamarin.Form
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Caliburn.Micro.WinAppSdk", "Caliburn.Micro.WinAppSdk\Caliburn.Micro.WinAppSdk.csproj", "{F885DD2B-D3B5-4263-AD5B-0B8FC79F6D60}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Caliburn.Micro.WinAppSdk.Test", "Caliburn.Micro.WinAppSdk.Test\Caliburn.Micro.WinAppSdk.Test.csproj", "{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Caliburn.Micros.WinAppSdk.TestApp", "Caliburn.Micros.WinAppSdk.TestApp\Caliburn.Micros.WinAppSdk.TestApp.csproj", "{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -143,30 +143,30 @@ Global
 		{F885DD2B-D3B5-4263-AD5B-0B8FC79F6D60}.Release|x64.Build.0 = Release|Any CPU
 		{F885DD2B-D3B5-4263-AD5B-0B8FC79F6D60}.Release|x86.ActiveCfg = Release|Any CPU
 		{F885DD2B-D3B5-4263-AD5B-0B8FC79F6D60}.Release|x86.Build.0 = Release|Any CPU
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Debug|Any CPU.ActiveCfg = Debug|x64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Debug|Any CPU.Build.0 = Debug|x64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Debug|Any CPU.Deploy.0 = Debug|x64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Debug|arm64.ActiveCfg = Debug|arm64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Debug|arm64.Build.0 = Debug|arm64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Debug|arm64.Deploy.0 = Debug|arm64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Debug|x64.ActiveCfg = Debug|x64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Debug|x64.Build.0 = Debug|x64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Debug|x64.Deploy.0 = Debug|x64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Debug|x86.ActiveCfg = Debug|x86
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Debug|x86.Build.0 = Debug|x86
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Debug|x86.Deploy.0 = Debug|x86
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Release|Any CPU.ActiveCfg = Release|x64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Release|Any CPU.Build.0 = Release|x64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Release|Any CPU.Deploy.0 = Release|x64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Release|arm64.ActiveCfg = Release|arm64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Release|arm64.Build.0 = Release|arm64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Release|arm64.Deploy.0 = Release|arm64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Release|x64.ActiveCfg = Release|x64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Release|x64.Build.0 = Release|x64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Release|x64.Deploy.0 = Release|x64
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Release|x86.ActiveCfg = Release|x86
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Release|x86.Build.0 = Release|x86
-		{635A6D6B-F15F-45FD-8335-D7BCDF8895AE}.Release|x86.Deploy.0 = Release|x86
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Debug|Any CPU.Build.0 = Debug|x64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Debug|Any CPU.Deploy.0 = Debug|x64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Debug|arm64.ActiveCfg = Debug|arm64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Debug|arm64.Build.0 = Debug|arm64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Debug|arm64.Deploy.0 = Debug|arm64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Debug|x64.ActiveCfg = Debug|x64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Debug|x64.Build.0 = Debug|x64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Debug|x64.Deploy.0 = Debug|x64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Debug|x86.ActiveCfg = Debug|x86
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Debug|x86.Build.0 = Debug|x86
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Debug|x86.Deploy.0 = Debug|x86
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Release|Any CPU.ActiveCfg = Release|x64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Release|Any CPU.Build.0 = Release|x64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Release|Any CPU.Deploy.0 = Release|x64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Release|arm64.ActiveCfg = Release|arm64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Release|arm64.Build.0 = Release|arm64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Release|arm64.Deploy.0 = Release|arm64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Release|x64.ActiveCfg = Release|x64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Release|x64.Build.0 = Release|x64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Release|x64.Deploy.0 = Release|x64
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Release|x86.ActiveCfg = Release|x86
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Release|x86.Build.0 = Release|x86
+		{56A6916B-28CF-47A1-9FE7-2C08727EB2FB}.Release|x86.Deploy.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
