+++ src/Caliburn.Micros.WinAppSdk.TestApp/App.xaml
@@ -1,16 +1,16 @@
-﻿<Application
-    x:Class="Caliburn.Micro.WinAppSdk.Test.App"
+﻿<cm:CaliburnApplication
+    x:Class="Caliburn.Micros.WinAppSdk.TestApp.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:cm="using:Caliburn.Micro"
-    xmlns:local="using:Caliburn.Micro.WinAppSdk.Test">
+    xmlns:local="using:Caliburn.Micros.WinAppSdk.TestApp">
     <Application.Resources>
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
-               <XamlControlsResources xmlns="using:Microsoft.UI.Xaml.Controls" />
-               <!--   Other merged dictionaries here -->
+                <XamlControlsResources xmlns="using:Microsoft.UI.Xaml.Controls" />
+                <!-- Other merged dictionaries here -->
             </ResourceDictionary.MergedDictionaries>
             <!-- Other app resources here -->
         </ResourceDictionary>
     </Application.Resources>
-</Application>
+</cm:CaliburnApplication>
