+++ src/Caliburn.Micros.WinAppSdk.TestApp/Assets/LockScreenLogo.scale-200.png
