+++ src/Caliburn.Micros.WinAppSdk.TestApp/Assets/SplashScreen.scale-200.png
