+++ src/Caliburn.Micros.WinAppSdk.TestApp/Assets/Square150x150Logo.scale-200.png
