+++ src/Caliburn.Micros.WinAppSdk.TestApp/Assets/Square44x44Logo.scale-200.png
