+++ src/Caliburn.Micros.WinAppSdk.TestApp/Assets/Square44x44Logo.targetsize-24_altform-unplated.png
