+++ src/Caliburn.Micros.WinAppSdk.TestApp/Assets/StoreLogo.png
