+++ src/Caliburn.Micros.WinAppSdk.TestApp/Assets/Wide310x150Logo.scale-200.png
