+++ src/Caliburn.Micros.WinAppSdk.TestApp/MainWindow.xaml
@@ -1,0 +1,13 @@
+﻿<Window
+    x:Class="Caliburn.Micros.WinAppSdk.TestApp.MainWindow"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Caliburn.Micros.WinAppSdk.TestApp"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <StackPanel>
+        <Frame x:Name="rootFrame" />
+    </StackPanel>
+</Window>
