+++ src/Caliburn.Micros.WinAppSdk.TestApp/ViewModels/HomeViewModel.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Caliburn.Micro;
+
+namespace Caliburn.Micros.WinAppSdk.TestApp.ViewModels
+{
+    public class HomeViewModel : Screen
+    {
+        private string _title;
+
+        public string Title
+        {
+            get => _title;
+            set
+            {
+                _title = value;
+                NotifyOfPropertyChange("Title");
+            }
+        }
+
+        public HomeViewModel()
+        {
+            Title = "Welcome to Caliburn Micro in UWP";
+        }
+
+        public void myButton()
+        {
+            Title = "Button Clicked";
+        }
+    }
+}
