+++ src/Caliburn.Micros.WinAppSdk.TestApp/Views/HomeView.xaml
@@ -1,14 +1,17 @@
 ﻿<UserControl
-    x:Class="Caliburn.Micro.WinAppSdk.Test.Views.HomeView"
+    x:Class="Caliburn.Micros.WinAppSdk.TestApp.Views.HomeView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Caliburn.Micro.WinAppSdk.Test.Views"
+    xmlns:local="using:Caliburn.Micros.WinAppSdk.TestApp.Views"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    Width="400"
+    Height="300"    
     mc:Ignorable="d">
+
     <StackPanel HorizontalAlignment="Center" VerticalAlignment="Center">
-        <TextBlock Text="{Binding Title}" />
-        <Button x:Name="myButton">Click Me</Button>
-        <TextBlock Text="{Binding Clicked}"></TextBlock>
+        <TextBlock Text="{Binding Path=Title}"></TextBlock>
+
+        <Button x:Name="myButton" >Click Me</Button>
     </StackPanel>
 </UserControl>
