+++ src/Caliburn.Micros.WinAppSdk.TestApp/Views/HomeView.xaml.cs
@@ -16,11 +16,8 @@ using Windows.Foundation.Collections;
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
 
-namespace Caliburn.Micro.WinAppSdk.Test.Views
+namespace Caliburn.Micros.WinAppSdk.TestApp.Views
 {
-    /// <summary>
-    /// An empty page that can be used on its own or navigated to within a Frame.
-    /// </summary>
     public sealed partial class HomeView : UserControl
     {
         public HomeView()
