+++ src/Caliburn.Micros.WinAppSdk.TestApp/app.manifest
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <assembly manifestVersion="1.0" xmlns="urn:schemas-microsoft-com:asm.v1">
-  <assemblyIdentity version="1.0.0.0" name="Caliburn.Micro.WinAppSdk.Test.app"/>
+  <assemblyIdentity version="1.0.0.0" name="Caliburn.Micros.WinAppSdk.TestApp.app"/>
 
   <application xmlns="urn:schemas-microsoft-com:asm.v3">
     <windowsSettings>
